pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.133.0",
    "google-genai>=1.64.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.21",
    "uvicorn>=0.41.0",
]
